v1.1.14 Update #14
Removed the updater from NMSCoordinates and created a separate whole new solution and repo for it.
