+++ NMSCoordinates.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.28803.352
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NMSCoordinates", "NMSCoordinates\NMSCoordinates.csproj", "{BEA99B18-A807-4178-B06F-D7FC7B923FE7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NMSCoordinatesUpdater", "NMSCoordinatesUpdater\NMSCoordinatesUpdater.csproj", "{186A98E2-410A-48F2-BF43-E6339ED53664}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{BEA99B18-A807-4178-B06F-D7FC7B923FE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BEA99B18-A807-4178-B06F-D7FC7B923FE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BEA99B18-A807-4178-B06F-D7FC7B923FE7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{186A98E2-410A-48F2-BF43-E6339ED53664}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{186A98E2-410A-48F2-BF43-E6339ED53664}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{186A98E2-410A-48F2-BF43-E6339ED53664}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{186A98E2-410A-48F2-BF43-E6339ED53664}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
