--- NMSCoordinatesUpdater/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Octokit" version="0.36.0" targetFramework="net472" />
-</packages>
